Tcs: Switch prices from f32 to f64 (#634)
To have more precision when storing user supplied prices and ensure that
the rounding that will occur will be tiny compared to the user-provided
precision.

+++ programs/mango-v4/tests/program_test/mango_client.rs
@@ -4138,9 +4138,9 @@ pub struct TokenConditionalSwapCreateInstruction {
     pub sell_mint: Pubkey,
     pub max_buy: u64,
     pub max_sell: u64,
-    pub price_lower_limit: f32,
-    pub price_upper_limit: f32,
-    pub price_premium_fraction: f32,
+    pub price_lower_limit: f64,
+    pub price_upper_limit: f64,
+    pub price_premium_fraction: f64,
     pub allow_creating_deposits: bool,
     pub allow_creating_borrows: bool,
 }
